% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElecPack-package.R
\docType{package}
\name{ElecPack-package}
\alias{ElecPack}
\alias{ElecPack-package}
\title{ElecPack: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
