% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WB_data.R
\docType{data}
\name{WB_data}
\alias{WB_data}
\title{世界银行-用电人口比例 1988-2020}
\format{
表格 with 271 rows and 37 variables:
\describe{
  \item{Country.Name}{国家英文全称}
  \item{Country.Code}{国家编号缩写}
  \item{Series.Name}{用电人口比例}
  \item{Series.Code}{ELC}
  \item{Year1988}{1988年用电人口比例}
  \item{......}{......}
  \item{Year2020}{2020年用电人口比例}
}
}
\source{
\url{https://databank.worldbank.org/reports.aspx?source=world-development-indicators#advancedDownloadOptions}
}
\usage{
WB_data
}
\description{
世界银行-用电人口比例 1988-2020
}
\keyword{datasets}
